Epics folder reorganisation (#195)
* split routing in separate folders

* split vault epics in separate folders

* renamed epics

* minor fix

* flattened epics folder structure
