+++ packages/dapp/src/epics/routing/index.js
@@ -1,1 +1,4 @@
-export { default } from './routing'
+import loginEpic from './login'
+import logoutEpic from './logout'
+
+export default [loginEpic, logoutEpic]
