+++ packages/dapp/src/epics/routing/logout.js
@@ -8,7 +8,7 @@ import { of } from 'rxjs/observable/of'
 import ROUTES from '../../constants/routes'
 import routerActions from '../../actions/router-actions'
 
-export const logoutEpic = (action$, store) => {
+const logoutEpic = (action$, store) => {
   const action$1 = action$
     .filter(action => action.type === LOCATION_CHANGE)
     .mergeMap(() => {
@@ -31,15 +31,4 @@ export const logoutEpic = (action$, store) => {
   return merge(action$1, action$2)
 }
 
-export const loginEpic = (action$, store) => {
-  return action$
-    .filter(action => action.type === actionTypes.LOGGED_IN)
-    .mergeMap(() => {
-      const state = store.getState()
-      return state.routing.location.pathname === ROUTES.LOGIN
-        ? of(routerActions.login())
-        : empty()
-    })
-}
-
-export default [logoutEpic, loginEpic]
+export default logoutEpic
