+++ packages/dapp/src/epics/vaults/index.js
@@ -1,1 +1,5 @@
-export { default } from './vaults'
+import fetchVaultEvents from './fetchVaultEvents'
+import registerVaults from './registerVaults'
+import watchVaultEvents from './watchVaultEvents'
+
+export default [fetchVaultEvents, registerVaults, watchVaultEvents]
