+++ packages/dapp/test/pages/preferences.js
@@ -20,7 +20,7 @@ module.exports = {
   },
 
   changeTimezoneValue(timezone) {
-    I.wait(0.2)
+    I.wait(0.4)
     I.click('div[id="1-toggle"]')
     I.waitForVisible('ul[id="1-menu-options"]')
     I.executeScript(() => {
